refactor:#8 이벤트 전파 문제 수정
