+++ src/hooks/usePlacesSearch.ts
@@ -1,5 +1,19 @@
 import { useEffect, useRef, useState } from 'react';
-import { Place } from '../components/common/Input/LocationInput';
+
+export interface Place {
+  address_name: string;
+  category_group_code: string;
+  category_group_name: string;
+  category_name: string;
+  distance: string;
+  id: string;
+  phone: string;
+  place_name: string;
+  place_url: string;
+  road_address_name: string;
+  x: string;
+  y: string;
+}
 
 const usePlacesSearch = () => {
   const [places, setPlaces] = useState<Place[]>([]);
